pyproject.toml
[tool.poetry]
name = "chatgpt"
version = "0.1.0"
description = ""
authors = ["antonzakhlebayeu"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
openai = "^1.30.5"
python-dotenv = "^1.0.1"
g4f = "^0.3.1.9"
python-telegram-bot = "^21.2"
sqlalchemy = "^2.0.30"
nest-asyncio = "^1.6.0"
selenium-wire = "^5.1.0"


[tool.poetry.group.dev.dependencies]
pylint = "^3.2.2"
black = "^24.4.2"
isort = "^5.13.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
